+++ src/app/example/configs/basic/date/form-config-date_zh-TW.constant.ts
@@ -1,0 +1,13 @@
+import { FormControlConfig } from 'projects/ng-dynamic-json-form/src/public-api';
+
+export const FORM_CONFIG_DATE_ZHTW: FormControlConfig = {
+  label: '日期',
+  formControlName: 'date',
+  type: 'date',
+  extra: {
+    date: {
+      selectTime: true,
+      outputFormat: 'yyyy/MM/dd HH:mm',
+    },
+  },
+};